handle install and update separately
